+++ src/Shrun/Configuration/Data/CommonLogging.hs
@@ -2,6 +2,7 @@
 
 module Shrun.Configuration.Data.CommonLogging
   ( -- * Types
+    Debug (..),
     CommonLoggingP (..),
     CommonLoggingArgs,
     CommonLoggingToml,
@@ -28,18 +29,51 @@ import Shrun.Configuration.Data.WithDisabled ((<.>?))
 import Shrun.Configuration.Default (Default (def))
 import Shrun.Prelude
 
+newtype Debug = MkDebug {unDebug :: Bool}
+  deriving stock (Eq, Show)
+
+instance
+  (k ~ An_Iso, a ~ Bool, b ~ Bool) =>
+  LabelOptic "unDebug" k Debug Debug a b
+  where
+  labelOptic = iso (\(MkDebug b) -> b) MkDebug
+  {-# INLINE labelOptic #-}
+
+instance Default Debug where
+  def = MkDebug False
+
+instance DecodeTOML Debug where
+  tomlDecoder = MkDebug <$> tomlDecoder
+
 -- | Holds command logging config.
 type CommonLoggingP :: ConfigPhase -> Type
-newtype CommonLoggingP p = MkCommonLoggingP
-  { -- | Whether to display command by (key) name or command.
+data CommonLoggingP p = MkCommonLoggingP
+  { -- | Whether debug logs are on.
+    debug :: ConfigPhaseF p Debug,
+    -- | Whether to display command by (key) name or command.
     keyHide :: ConfigPhaseF p KeyHideSwitch
   }
 
 instance
-  (k ~ An_Iso, a ~ ConfigPhaseF p KeyHideSwitch, b ~ ConfigPhaseF p KeyHideSwitch) =>
+  (k ~ A_Lens, a ~ ConfigPhaseF p Debug, b ~ ConfigPhaseF p Debug) =>
+  LabelOptic "debug" k (CommonLoggingP p) (CommonLoggingP p) a b
+  where
+  labelOptic =
+    lensVL $ \f (MkCommonLoggingP a1 a2) ->
+      fmap
+        (\b -> MkCommonLoggingP b a2)
+        (f a1)
+  {-# INLINE labelOptic #-}
+
+instance
+  (k ~ A_Lens, a ~ ConfigPhaseF p KeyHideSwitch, b ~ ConfigPhaseF p KeyHideSwitch) =>
   LabelOptic "keyHide" k (CommonLoggingP p) (CommonLoggingP p) a b
   where
-  labelOptic = iso (\(MkCommonLoggingP kh) -> kh) MkCommonLoggingP
+  labelOptic =
+    lensVL $ \f (MkCommonLoggingP a1 a2) ->
+      fmap
+        (\b -> MkCommonLoggingP a1 b)
+        (f a2)
   {-# INLINE labelOptic #-}
 
 type CommonLoggingArgs = CommonLoggingP ConfigPhaseArgs
@@ -63,10 +97,10 @@ deriving stock instance Eq (CommonLoggingP ConfigPhaseMerged)
 deriving stock instance Show (CommonLoggingP ConfigPhaseMerged)
 
 instance
-  (Default (ConfigPhaseF p KeyHideSwitch)) =>
+  (Default (ConfigPhaseF p Debug), Default (ConfigPhaseF p KeyHideSwitch)) =>
   Default (CommonLoggingP p)
   where
-  def = MkCommonLoggingP def
+  def = MkCommonLoggingP def def
 
 -- | Merges args and toml configs.
 mergeCommonLogging ::
@@ -75,7 +109,9 @@ mergeCommonLogging ::
   CommonLoggingMerged
 mergeCommonLogging args mToml =
   MkCommonLoggingP
-    { keyHide =
+    { debug =
+        (args ^. #debug) <.>? (toml ^. #debug),
+      keyHide =
         (args ^. #keyHide) <.>? (toml ^. #keyHide)
     }
   where
@@ -84,7 +120,8 @@ mergeCommonLogging args mToml =
 instance DecodeTOML CommonLoggingToml where
   tomlDecoder =
     MkCommonLoggingP
-      <$> decodeKeyHideSwitch
+      <$> getFieldOptWith tomlDecoder "debug"
+      <*> decodeKeyHideSwitch
 
 decodeKeyHideSwitch :: Decoder (Maybe KeyHideSwitch)
 decodeKeyHideSwitch = getFieldOptWith tomlDecoder "key-hide"
@@ -93,5 +130,6 @@ decodeKeyHideSwitch = getFieldOptWith tomlDecoder "key-hide"
 toEnv :: CommonLoggingMerged -> CommonLoggingEnv
 toEnv merged =
   MkCommonLoggingP
-    { keyHide = merged ^. #keyHide
+    { debug = merged ^. #debug,
+      keyHide = merged ^. #keyHide
     }